Epi framing v2 (#7)
* We have a prelude now

* This doesn't belong in main

* Adding scenario-builders and a generic

* Updated packages

* Added the compounding of multiple
parameters for the same outcome.

* Unless I'm using the const-generic
then there is no need to use it..

* Documentation

* Tried to make a default generic so not necessarily
having to add a population
tied herd-size.

* comment

* Task

* Minor formatting

* Something strange is happening about
this. I think it is incompatibility across
the platform. These files help;

* this could be a derive-macro. Cannot figure that out right now.

* Added an example where we change the world container directly

* Refactored such that the cattle_populations module can be phased out
Note that the PhantomData part needs to not be
displayed, maybe only in Debug, this a Display-impl
should be derived for these things.

* Using resource-scope instead of
accessing things in sequence.

* `cargo fmt`

* Refactored scenario time stuff
Added a test case (and learned more about
this stuff)

* `cargo fmt`

* Removed deprecated module

* Minor edit

* Minor edit

* Minor change to printing.

* clippy

* `cargo fmt`

* Readjusted output

* Plan

* Branch needs to be specified

* Added extra lints

* Prelude as a file

* added bacon,
but this needs more flags